fix(web): give the export queue five states and its mutations their own errors
771 lines in one route file, hand-rolling the loading/error/empty triple that
useAsyncResource and ResourceState exist to hold. Extracted to
ResearchExportsScreen, leaving the guard-only route file the migrated routes
already use.

The state work, in order of how much it mattered:

  - one `error` covered the queue read and all five mutations, so a download
    that failed rendered as "we couldn't load this view" above a list that had
    loaded perfectly well, offering to reload the queue as the recovery. The
    read's failure is ResourceState's business now; the operator's action
    failing is a different sentence and keeps its own state.
  - the 10s poll called loadExports({ background: true }), a hand-rolled
    approximation of isRefreshing. It is a refresh() now, so a poll that times
    out on clinic wifi leaves the last good queue on screen under a stale
    banner instead of clearing it.
  - the empty case was a dashed div with no heading and no icon; the error case
    was a red div with no retry at all.
  - the read had no AbortSignal and no request-id guard, so a slow response
    could overwrite a newer one.

Also: the local hasPermission was a byte-copy of lib/ops.ts:208 and now imports
it; the pack-contents list wrapped filenames in markdown backticks, which
rendered as literal backtick characters; the date range now marks the inputs
aria-invalid and points them at the message rather than leaving it floating;
and the two cards drop `shadow-lg shadow-black/5`, which MASTER.md section 7
does not allow on a clinical surface.

Refs #22, #82
